+++ packages/utils/src/build-scripts/core/core.pre-build.ts
@@ -1,4 +1,3 @@
-import { generateIndex } from './core-generate-index'
 import { generateMetadataTs } from './core-generate-metadata'
 import { generateSvgModule } from './core-generate-svgs-module'
 import { optimizeSVGs } from './core-optimize-svgs'
@@ -6,7 +5,6 @@ import { optimizeSVGs } from './core-optimize-svgs'
 function preBuild() {
   optimizeSVGs()
   generateSvgModule()
-  generateIndex()
   generateMetadataTs()
 }