+++ packages/utils/src/scripts/gecko/merge-custom.ts
@@ -3,7 +3,7 @@ import path from 'path'
 import customTokens from './custom-tokens.json'
 import geckoCoins from './gecko-coins.json'
 import getCoinByID from './get-coin-by-id'
-import { ITokenMetadata, TVariant } from '../../types'
+import { ITokenMetadata, TVariant } from '@web3icons/common'
 import { SVG_TOKENS_SRC_DIR, TOKENS_METADATA_PATH } from '../../constants'
 import prettier from 'prettier'