+++ packages/utils/src/utils/append-to-metadata.ts
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import { TOKENS_METADATA_PATH } from '../constants'
-import { ITokenMetadata } from '../types'
+import { ITokenMetadata } from '@web3icons/common'
 
 export const appendToMetadataJson = (coin: ITokenMetadata) => {
   let fileContent = fs.existsSync(TOKENS_METADATA_PATH)