+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web3icons/core
 
+## 3.6.2-alpha.0
+
+### Patch Changes
+
+- common package that holds types and metadata
+
 ## 3.6.1
 
 ### Patch Changes