+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @web3icons/react
 
+## 3.6.2-alpha.0
+
+### Patch Changes
+
+- common package that holds types and metadata
+- Updated dependencies
+  - @web3icons/core@3.6.2-alpha.0
+
 ## 3.6.1
 
 ### Patch Changes